refactor

+++ src/test/java/com/sfhuskie/plugins/githooks/io/TestScriptMaker.java
@@ -44,7 +44,7 @@ public class TestScriptMaker {
     @Test
     public void getCommandToAdd() throws IOException {
         File script = new File(MojoSettings.userDir+"/"+MojoSettings.PRECOMMIT);
-        assertEquals("bash "+script.getAbsolutePath(), ScriptMaker.getCommandToAdd(script));
+        assertEquals("bash "+script.getCanonicalPath(), MojoSettings.getCommandToAdd(script));
     }
     @Test
     public void getCommandsPrecommit() throws IOException {
@@ -53,18 +53,20 @@ public class TestScriptMaker {
         File precommit = new File(MojoSettings.userDir+"/"+MojoSettings.PRECOMMIT);
         File prepush = new File(MojoSettings.userDir+"/"+MojoSettings.PREPUSH);
         List<String> cmds = sm.getCommands(precommit);
-        assertTrue(cmds.contains(ScriptMaker.getCommandToAdd(precommit)));
+        assertTrue(cmds.contains(MojoSettings.getCommandToAdd(precommit)));
         assertTrue(cmds.get(0).equals("#!/bin/sh"));
         assertTrue(!cmds.contains(prepush.getAbsolutePath()));
     }
     @Test
-    public void generate() throws IOException {
+    public void generateScripts() throws IOException {
         hooks.add(MojoSettings.PRECOMMIT);
         ScriptMaker sm = new ScriptMaker(tools, hooks);
         File destDir = new File(MojoSettings.userDir+"/target/bin");
         File precommit = new File(MojoSettings.userDir+"/target/bin/"+MojoSettings.PRECOMMIT);
-        FileUtils.delete(precommit);
-        sm.generate(destDir);
+        if (precommit.exists()) {
+            FileUtils.delete(precommit);
+        }
+        sm.generateScripts(destDir);
         assertTrue(precommit.exists());
         List<String> cmds = sm.getCommands(precommit);
         assertEquals(cmds, FileIO.getLines(precommit));
